+++ Geoscape/RecruitInfo/RecruitInfo.cs
@@ -34,9 +34,11 @@ namespace Sheepy.PhoenixPt.RecruitInfo {
    public class Mod : ZyMod {
       internal static ModConfig Config;
 
-      public static void Init () => new Mod().MainMod();
+      public static void Init () => new Mod().GeoscapeMod();
 
-      public void MainMod ( Func< string, object, object > api = null ) {
+      public void MainMod ( Func< string, object, object > api ) => GeoscapeMod( api );
+
+      public void GeoscapeMod ( Func< string, object, object > api = null ) {
          SetApi( api, out Config );
          if ( Config.Grafts.Enabled ) {
             var sharedData = GameUtl.GameComponent<SharedData>();