+++ Tactical/DeafTheDead/changelog.md
@@ -1,5 +1,9 @@
 Changelog of Deaf the Dead, a Phoenix Point Mod by Sheepy
 
+# Version ?,  in development
+
+* Support Modnix 3 lazy loading
+
 # Version 1, 2020-04-13
 
 * Make sure dead unit cannot hear enemies.