+++ Tactical/MitigateShred/changelog.md
@@ -2,7 +2,7 @@ Changelog of Mitigate Shred, a Phoenix Point Mod by Sheepy
 
 # Version 2, in development
 
-* New: Updated to Modnix and made configurable.
+* New: Updated to Modnix and made configurable and lazy loaded.
 
 # Version 1, 2019-12-29